refactor models module into package
- still have some local problems in PageOperationMixin, but
- at least unit tests pass
