+++ models/schema_data_index.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+import logging
+from google.appengine.ext import ndb
+
+
+logging.getLogger().setLevel(logging.DEBUG)
+
+class SchemaDataIndex(ndb.Model):
+    title = ndb.StringProperty()
+    name = ndb.StringProperty()
+    value = ndb.StringProperty()
+
